Fixes DanThrane/tekvideo.sdu.dk#111
This (visual) bug was present, due to slight differences in how the
polyfill works and how the real thing works. The problem would occur
when we have external code working inside of the shadow DOM. Normally
code working inside of the component would add the correct classes to
all elements, and thus getting the correct classes added for the style
scoping. However since this code was completely external it wouldn't
happen.

A small function called "addExtSheet" has been added to
"tekvideo-math-field.html". This function will detect if we're using
the polyfill or not, depending on that it will include the external
stylesheet either directly to the document head (protected against
duplicates) or to component itself. If this bug re-apears elsewhere
the same solution should be used.